# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Roblox Rewards Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterRequest {
  username: string;
  email: string;
  password: string;
  inviteCode?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
  email: string;
  points: number;
  robuxBalance: number;
  inviteCode: string;
  totalInvites: number;
  weeklyInvites: number;
  level: number;
  isVerified: boolean;
  lastDailyReward?: string | null;
  createdAt?: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface DailyRewardResponse {
  pointsEarned: number;
  newTotal: number;
}

export interface InviteInfo {
  inviteCode: string;
  inviteLink: string;
  totalInvites: number;
  pointsFromInvites: number;
  nextMilestone?: number | null;
  nextMilestonePoints?: number | null;
}

export interface Reward {
  id: number;
  nameAr: string;
  pointsCost: number;
  robuxValue: number;
  requiredLevel: number;
  isLocked: boolean;
  descriptionAr?: string | null;
}

export interface RedeemRequest {
  rewardId: number;
}

export interface LeaderboardEntry {
  rank: number;
  username: string;
  totalInvites: number;
  points: number;
}

export interface LeaderboardResponse {
  entries: LeaderboardEntry[];
  resetInDays: number;
  currentUserRank?: number | null;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  earn: "earn",
  spend: "spend",
  withdraw: "withdraw",
  invite_bonus: "invite_bonus",
  daily_reward: "daily_reward",
} as const;

export interface Transaction {
  id: number;
  type: TransactionType;
  amount: number;
  description: string;
  createdAt: string;
}

export type WithdrawalRequestPaymentMethod =
  (typeof WithdrawalRequestPaymentMethod)[keyof typeof WithdrawalRequestPaymentMethod];

export const WithdrawalRequestPaymentMethod = {
  visa: "visa",
  vodafone: "vodafone",
  orange: "orange",
  etisalat: "etisalat",
} as const;

export interface WithdrawalRequest {
  amountRobux: number;
  paymentMethod: WithdrawalRequestPaymentMethod;
  paymentDetails: string;
}

export type WithdrawalStatus =
  (typeof WithdrawalStatus)[keyof typeof WithdrawalStatus];

export const WithdrawalStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  rejected: "rejected",
} as const;

export interface Withdrawal {
  id: number;
  amountRobux: number;
  paymentMethod: string;
  status: WithdrawalStatus;
  rejectionReason?: string | null;
  requestedAt: string;
  processedAt?: string | null;
}