# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Roblox Rewards Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/daily-reward:
    post:
      operationId: claimDailyReward
      tags: [users]
      summary: Claim daily reward
      responses:
        "200":
          description: Daily reward claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyRewardResponse"
        "400":
          description: Already claimed today
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        inviteCode:
          type: string
      required:
        - username
        - email
        - password
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        points:
          type: integer
        robuxBalance:
          type: integer
        inviteCode:
          type: string
        totalInvites:
          type: integer
        weeklyInvites:
          type: integer
        level:
          type: integer
        isVerified:
          type: boolean
        lastDailyReward:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - username
        - email
        - points
        - robuxBalance
        - inviteCode
        - totalInvites
        - weeklyInvites
        - level
        - isVerified
    DailyRewardResponse:
      type: object
      properties:
        pointsEarned:
          type: integer
        newTotal:
          type: integer
      required:
        - pointsEarned
        - newTotal
...