# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Roblox Rewards Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  email: string;
  points: number;
  robuxBalance: number;
  inviteCode: string;
  totalInvites: number;
  weeklyInvites: number;
  level: number;
  isVerified: boolean;
  lastDailyReward?: string | null;
  createdAt?: string;
}